# docs/openapi.json
{
	"openapi": "3.1.0",
	"info": { "title": "SocialAnimal REST", "version": "1.0.0" },
	"components": {
		"securitySchemes": {
			"bearerAuth": { "type": "http", "scheme": "bearer" }
		}
	},
	"paths": {
...
		"/api/v1/calendars": {
			"get": {
				"operationId": "calendars",
				"parameters": [
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"minimum": 1,
							"maximum": 500,
							"default": 100
						}
					},
					{
						"name": "cursor",
						"in": "query",
						"required": false,
						"schema": { "type": "string" }
					}
				],
				"security": [{ "bearerAuth": [] }],
				"responses": {
					"200": {
						"description": "200",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"items": {
											"type": "array",
											"items": {
												"type": "object",
												"additionalProperties": false,
												"properties": {
													"id": { "type": "string" },
													"name": {
														"type": "string"
													},
													"type": {
														"type": "string"
													},
													"connectionId": {
														"type": [
															"null",
															"string"
														]
													},
													"remoteId": {
														"type": [
															"null",
															"string"
														]
													},
													"syncInterval": {
														"type": "integer"
													},
													"eventCount": {
														"type": "integer"
													},
													"lastSync": {
														"type": [
															"null",
															"string"
														]
													},
													"lastAttempt": {
														"type": [
															"null",
															"string"
														]
													},
													"lastSuccess": {
														"type": [
															"null",
															"string"
														]
													},
													"lastError": {
														"type": [
															"null",
															"string"
														]
													}
												},
												"required": [
													"id",
													"name",
													"type",
													"syncInterval"
												]
											}
										},
										"nextCursor": {
											"type": ["null", "string"]
										}
									},
									"required": ["items", "nextCursor"]
								}
							}
						}
					},
					"400": {
						"description": "400",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"401": {
						"description": "401",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"403": {
						"description": "403",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"404": {
						"description": "404",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"409": {
						"description": "409",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"422": {
						"description": "422",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"500": {
						"description": "500",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"502": {
						"description": "502",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"503": {
						"description": "503",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					}
				}
			}
		},
...
		"/api/v1/calendars/{id}/grants/{userId}": {
			"put": {
				"operationId": "setGrant",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"minLength": 1,
							"maxLength": 2048
						}
					},
					{
						"name": "userId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"minLength": 1,
							"maxLength": 2048
						}
					}
				],
				"security": [{ "bearerAuth": [] }],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"additionalProperties": false,
								"properties": {
									"permission": {
										"type": "string",
										"enum": ["busy", "titles", "full"]
									}
								},
								"required": ["permission"]
							}
						}
					}
				},
				"responses": {
					"204": { "description": "204" },
					"400": {
						"description": "400",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"401": {
						"description": "401",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"403": {
						"description": "403",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"404": {
						"description": "404",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"409": {
						"description": "409",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"422": {
						"description": "422",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"500": {
						"description": "500",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"502": {
						"description": "502",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"503": {
						"description": "503",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "removeGrant",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"minLength": 1,
							"maxLength": 2048
						}
					},
					{
						"name": "userId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"minLength": 1,
							"maxLength": 2048
						}
					}
				],
				"security": [{ "bearerAuth": [] }],
				"responses": {
					"204": { "description": "204" },
					"400": {
						"description": "400",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"401": {
						"description": "401",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"403": {
						"description": "403",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"404": {
						"description": "404",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"409": {
						"description": "409",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"422": {
						"description": "422",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"500": {
						"description": "500",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"502": {
						"description": "502",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"503": {
						"description": "503",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					}
				}
			}
		},
...
		"/api/v1/subscriptions/{id}": {
			"delete": {
				"operationId": "revokeSubscription",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"minLength": 1,
							"maxLength": 2048
						}
					}
				],
				"security": [{ "bearerAuth": [] }],
				"responses": {
					"204": { "description": "204" },
					"400": {
						"description": "400",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"401": {
						"description": "401",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"403": {
						"description": "403",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"404": {
						"description": "404",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"409": {
						"description": "409",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"422": {
						"description": "422",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"500": {
						"description": "500",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"502": {
						"description": "502",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"503": {
						"description": "503",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					}
				}
			}
		}
	}
}